refactor: single component header and dedicated back button
