+++ frontend/svelte/src/lib/components/common/Back.svelte
@@ -1,0 +1,24 @@
+<script lang="ts">
+  import IconBackIosNew from "../../icons/IconBackIosNew.svelte";
+  import { i18n } from "../../stores/i18n";
+  import Tooltip from "../ui/Tooltip.svelte";
+  import { createEventDispatcher } from "svelte";
+
+  const dispatch = createEventDispatcher();
+</script>
+
+<Tooltip id="back" text={$i18n.core.back} slot="start">
+  <button
+    class="back icon-only"
+    on:click|stopPropagation={() => dispatch("nnsBack")}
+    aria-label={$i18n.core.back}><IconBackIosNew /></button
+  >
+</Tooltip>
+
+<style lang="scss">
+  @use "../../themes/mixins/header";
+
+  .back {
+    @include header.button(--brand-sea-buckthorn);
+  }
+</style>
