+++ frontend/svelte/src/lib/components/common/Header.svelte
@@ -1,14 +1,13 @@
 <script lang="ts">
   import Logout from "./Logout.svelte";
   import { triggerDebugReport } from "../../utils/dev.utils";
-  import Menu from "./Menu.svelte";
   import Header from "../ui/Header.svelte";
   import Toolbar from "../ui/Toolbar.svelte";
 </script>
 
 <Header>
   <Toolbar>
-    <Menu slot="start" />
+    <svelte:fragment slot="start"><slot name="start" /></svelte:fragment>
     <h4 use:triggerDebugReport><slot /></h4>
     <Logout slot="end" />
   </Toolbar>
