+++ frontend/svelte/src/lib/components/common/Layout.svelte
@@ -1,14 +1,16 @@
 <script lang="ts">
   import Footer from "./Footer.svelte";
-  import LayoutHeader from "./LayoutHeader.svelte";
+  import Header from "./Header.svelte";
   import Banner from "./Banner.svelte";
+  import Menu from "./Menu.svelte";
 </script>
 
 <Banner />
 
-<LayoutHeader>
+<Header>
+  <Menu slot="start" />
   <svelte:fragment><slot name="header" /></svelte:fragment>
-</LayoutHeader>
+</Header>
 
 <main>
   <slot />
